created auth route and implemenred admin and student login in authController
