+++ server/index.js
@@ -5,12 +5,14 @@ dotenv.config();
 import "./config/dbConfig.js";
 import studentRoute from "./routes/studentRoute.js";
 import { errorMiddleware } from "./middlewares/errorMiddleware.js";
+import authRoute from "./routes/authRoute.js";
 
 app.use(express.json());
 
 const apiBasePath = "/api/v1/";
 
 app.use(`${apiBasePath}student/`, studentRoute);
+app.use(`${apiBasePath}auth/`, authRoute);
 
 app.get("/", (req, res) => {
   res.status(200).json({ message: "Welcome to GradEase" });
