+++ server/services/studentService.js
@@ -3,38 +3,6 @@ import jwt from "jsonwebtoken";
 import StudentModel from "../models/studentModel.js";
 import ResponseError from "../utils/RestResponseError.js";
 
-export async function loginStudent(email, password) {
-  const existingStudent = await StudentModel.findOne({ email });
-
-  if (!existingStudent) {
-    throw new ResponseError(404, "Invalid email, email is not registered");
-  }
-
-  const isPasswordCorrect = await bcrypt.compare(
-    password,
-    existingStudent.password
-  );
-
-  if (!isPasswordCorrect) {
-    throw new ResponseError(
-      404,
-      "Invalid password, please enter the correct password"
-    );
-  }
-
-  const tokenObject = {
-    _id: existingStudent._id,
-    fullName: existingStudent.fullName,
-    email: existingStudent.email,
-  };
-
-  const jwtToken = jwt.sign(tokenObject, process.env.JWT_STUDENT_SECRET, {
-    expiresIn: "12h",
-  });
-
-  return jwtToken;
-}
-
 export async function registerStudent(studentModel) {
   const existingStudent = await StudentModel.findOne({
     email: studentModel.email,
